Refactor ToC color picker component into reusable palette picker
